Test noise refresh formatter with native PRG material
